+++ libs/material/src/lib/dynamic-forms.module.spec.ts
@@ -7,6 +7,7 @@ import {
   DynamicFormBuilder,
   DynamicFormComponentFactory,
   DynamicFormConfigService,
+  DynamicFormDateConverter,
   DynamicFormEvaluationBuilder,
   DynamicFormExpressionBuilder,
   DynamicFormIdBuilder,
@@ -143,7 +144,15 @@ describe('MatDynamicFormsModule', () => {
 
         it('provides DynamicFormValidationService', inject([DynamicFormValidationService], (service: DynamicFormValidationService) => {
           expect(service).toBeTruthy();
-          expect(service.validationConfig).toEqual({ ...dynamicFormValidationConfig, libraryName: matDynamicFormLibrary.name });
+          expect(service.validationConfig).toEqual({
+            ...dynamicFormValidationConfig,
+            aliases: {
+              ...dynamicFormValidationConfig.aliases,
+              matDatepickerMin: 'minDate',
+              matDatepickerMax: 'maxDate',
+            },
+            libraryName: matDynamicFormLibrary.name,
+          });
         }));
 
         it('provides DynamicFormComponentFactory', inject([DynamicFormComponentFactory], (service: DynamicFormComponentFactory) => {
@@ -154,6 +163,10 @@ describe('MatDynamicFormsModule', () => {
           expect(service).toBeTruthy();
           expect(service.handlers.length).toBe(25);
         }));
+
+        it('does not provide DynamicFormDateConverter', () => {
+          expect(() => TestBed.inject(DynamicFormDateConverter)).toThrowError(/NullInjectorError/);
+        });
       });
     });
   });