+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 18.0.0-rc.0 (2024-05-25)
+
+### Features
+
+* **core:** update to angular 18
+* **bootstrap:** update to angular 18
+* **bootstrap:** support of min date and max date for datepicker
+* **material:** update to angular and angular material 18
+* **material:** support of min date and max date for datepicker
+* **markdown:** update to angular 18
+
 ## 18.0.0-next.0 (2024-05-01)
 
 ### Features